Carry bearings in double, not float
The camera refactor stored normalised coordinates in cv::Point2f, rounding the
bearing before it reached the estimator; previously it was computed and
consumed in double. Image mode should have been behaviour-preserving and was
not: divergence from the baseline started at 2e-9 m on the first frame and
amplified to 6.5 cm max / 4.4 cm mean over the 300 s replay, because a
nonlinear sliding-window solve compounds a perturbation in every residual.

liftProjective now outputs Point2d, the image-mode normalisation casts to
double before dividing, and VinsBackend takes Point2d. The fisheye path also
promotes pixels to double before cv::fisheye::undistortPoints, which matches
output type to input -- float pixels in would have rounded the result even with
a double output vector.

Pixels stay float (that is what the detectors produce); only the bearing is
double. Documented at the interface with the measured numbers, since this is
exactly the kind of asymmetry someone later "simplifies" away.

Verified: image mode is now bit-identical to the 247.44 m baseline. Worth
noting 6.5 cm on a 4321 m trajectory is 0.0015% -- invisible in ATE, so only a
bit-identity check could catch it.

+++ odometry/OdometryPipeline.cpp
@@ -97,8 +97,8 @@ void OdometryPipeline::processFrame(DeviceBuffer& frame,
         const CameraIntrinsics& K = config.intrinsics;
         norm_.resize(fr.points2D.size());
         for (size_t i = 0; i < fr.points2D.size(); ++i) {
-            norm_[i].x = (fr.points2D[i].x - K.cx) / K.fx;
-            norm_[i].y = (fr.points2D[i].y - K.cy) / K.fy;
+            norm_[i].x = (static_cast<double>(fr.points2D[i].x) - K.cx) / K.fx;
+            norm_[i].y = (static_cast<double>(fr.points2D[i].y) - K.cy) / K.fy;
         }
     }
