+++ odometry/camera/ICameraModel.h
@@ -46,6 +46,12 @@ public:
     // (x/z, y/z). This is the Points-mode counterpart of undistortImage: it
     // applies the same lens model to a sparse set of points instead of to every
     // pixel. `out` is resized to match `px`.
+    //
+    // Output is double, not float, and deliberately so. Pixels are float
+    // because that is what the detectors produce, but the bearing feeds a
+    // nonlinear solve: rounding it to float shifts the estimate by ~2e-9 m on
+    // the first frame and 6.5 cm by the end of a 300 s replay. Measured, not
+    // hypothetical -- it is what the image-mode regression caught.
     virtual void liftProjective(const std::vector<cv::Point2f>& px,
-                                std::vector<cv::Point2f>& out) const = 0;
+                                std::vector<cv::Point2d>& out) const = 0;
 };
