+++ odometry/camera/KannalaBrandtCamera.h
@@ -14,7 +14,7 @@ public:
     cv::Size outputSize() const override { return output_size_; }
     cv::Mat undistortImage(const cv::Mat& raw) const override;
     void liftProjective(const std::vector<cv::Point2f>& px,
-                        std::vector<cv::Point2f>& out) const override;
+                        std::vector<cv::Point2d>& out) const override;
 
 private:
     CameraIntrinsics intrinsics_;
